use engine riscv64 sysroot, install sysroot in prepopulate step on macos
we no longer need to patch the engine sources to be able to download a
linux sysroot on macos.

So now we can install the sysroot in the prepopulate step for macos too.

All sysroot together are roughly 100M when compressed, so they don't
contribute much to overall cache size.